Error Handling Trips api

+++ client/src/api/trips-api.js
@@ -1,43 +1,62 @@
 import * as fetcher from "./fetcher";
 
-const BASE_URL = 'http://localhost:3000/api/trips';
+const BASE_URL = "http://localhost:3000/api/trips";
 
-async function getAllTrips () {
-    const result = await fetcher.get(BASE_URL);
-    const trips = Object.values(result);
-
-    return trips;
-
+async function getAllTrips() {
+    try {
+        const result = await fetcher.get(BASE_URL);
+        const trips = Object.values(result);
+        return trips;
+    } catch (error) {
+        throw new Error("Failed to fetch trips.");
+    }
 }
 
-async function getOneTrip (tripId) {
-    const trip = await fetcher.get(`${BASE_URL}/${tripId}/details`);
-
-    return trip;
-    
+async function getOneTrip(tripId) {
+    try {
+        if (!tripId) {
+            throw new Error("Trip Id is required");
+        }
+        const trip = await fetcher.get(`${BASE_URL}/${tripId}/details`);
+        return trip;
+    } catch (error) {
+        throw new Error("Failed to fetch trip details.");
+    }
 }
 
-async function createTrip (tripData) {
-    const createdTrip = await fetcher.post(`${BASE_URL}/create`, tripData);
-
-    return createdTrip;
-    
+async function createTrip(tripData) {
+    try {
+        const createdTrip = await fetcher.post(`${BASE_URL}/create`, tripData);
+        return createdTrip;
+    } catch (error) {
+        throw new Error("Failed to create trip.");
+    }
 }
 
 //Check, if tripId needed
-async function updateTrip (tripData) {
-    const updatedTrip = await fetcher.put(`${BASE_URL}/${tripId}/edit`, tripData);
-
-    return updatedTrip;
-    
+async function updateTrip(tripId, tripData) {
+    try {
+        if (!tripId) {
+            throw new Error("Trip Id is required.");
+        }
+        const updatedTrip = await fetcher.put(`${BASE_URL}/${tripId}/edit`, tripData);
+        return updatedTrip;
+    } catch (error) {
+        throw new Error("Failed to update trip.");
+    }
 }
 
 //Check, if tripId needed
-async function deleteTrip () {
-    const deletedTrip = await fetcher.del(`${BASE_URL}/${tripId}/delete`);
-
-    return deletedTrip;
-    
+async function deleteTrip() {
+    try {
+        if (!tripId) {
+            throw new Error("Trip Id is required.");
+        }
+        const deletedTrip = await fetcher.del(`${BASE_URL}/${tripId}/delete`);
+        return deletedTrip;
+    } catch (error) {
+        throw new Error("Failed to delete trip.");
+    }
 }
 
 export default {
@@ -45,5 +64,5 @@ export default {
     getOneTrip,
     createTrip,
     updateTrip,
-    deleteTrip
+    deleteTrip,
 };
